backend/requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0

# PDF Processing
pypdf==4.0.1
pdfplumber==0.10.4
pymupdf==1.23.26

# Embeddings & ML
sentence-transformers==2.3.1
torch>=2.1.0
transformers==4.37.2

# Vector Store - FAISS with GPU support
faiss-cpu==1.7.4
# For GPU: pip install faiss-gpu (requires CUDA)

# BM25 for hybrid retrieval
rank-bm25==0.2.2

# Cross-encoder for reranking
# Already included in sentence-transformers

# LLM Client
openai==1.12.0
httpx==0.26.0

# Async support
aiofiles==23.2.1
aiohttp==3.9.3

# Utilities
python-dotenv==1.0.1
tqdm==4.66.2
numpy==1.26.4
pandas==2.2.0

# Logging and monitoring
loguru==0.7.2
rich==13.7.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4

# Streamlit UI
streamlit==1.31.0
